[MM-68428] Consolidate webapp clients under webapp/src/clients/ folder (#1162)
