+++ webapp/src/components/incoming_calls/call_incoming_condensed.tsx
@@ -7,6 +7,7 @@ import React, {useEffect, useRef, useState} from 'react';
 import {OverlayTrigger, Tooltip} from 'react-bootstrap';
 import {FormattedMessage, useIntl} from 'react-intl';
 import {useSelector} from 'react-redux';
+import RestClient from 'src/clients/rest';
 import Avatar from 'src/components/avatar/avatar';
 import {Button} from 'src/components/buttons';
 import CompassIcon from 'src/components/icons/compassIcon';
@@ -16,7 +17,6 @@ import {
     useOnChannelLinkClick,
     useRingingAndNotification,
 } from 'src/components/incoming_calls/hooks';
-import RestClient from 'src/rest_client';
 import {ChannelType, IncomingCallNotification} from 'src/types/types';
 import styled, {css} from 'styled-components';
