+++ webapp/src/utils.ts
@@ -18,9 +18,9 @@ import {getCurrentRelativeTeamUrl, getCurrentTeamId, getTeam} from 'mattermost-r
 import {getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
 import {IntlShape} from 'react-intl';
 import {parseSemVer} from 'semver-parser';
-import CallsClient from 'src/client';
+import CallsClient from 'src/clients/calls';
+import RestClient from 'src/clients/rest';
 import {STORAGE_CALLS_SHARE_AUDIO_WITH_SCREEN} from 'src/constants';
-import RestClient from 'src/rest_client';
 import {DesktopMessage} from 'src/types/types';
 import {notificationSounds} from 'src/webapp_globals';
